Tebex Checkout Flow and Basket Management Integration (#1)
* Improved modal behavior and laid ground work for Tebex request triggering in AuthForms component

This commit enhances the behavior of the AuthForms component within a modal. The implementation ensures that the store request is triggered only when the modal is open and the user is authorized. Additionally, the code has been accommodated to prevent unnecessary store request triggers, particularly on logout. We've added the useEffect dependencies to accurately reflect the conditions for triggering the store request and added appropriate logging for clarity and debugging purposes.

* data integrity cleanup and added remix-utils, is-ip and remix-dev.

* stable version of create basket before we refactor and replace current implementation position for add package.

* setup basketId cookie and integrated in loader

* setup add to cart

* minimized cookie storage size > 4kb

fixed issue where html in description of received object in adding package to existing basket was over 4kb which prevented successful storage.

* setup subsequent request, enforced cookie typing and added checkout url.

made it so if user doesnt have a basket and we create one, we then add the premium membership afterwards. if user already had a basket, we simply add the package.

* eslint and prettier

todo: tailwind ruling and full stylesheet conversion (in different branch)

* Implement Tebex checkout integration with dynamic basket ID support

This commit concludes our foundational introduction to the integration of the Tebex checkout library into our platform. The key feature implemented includes the ability to dynamically trigger the Tebex checkout modal based on a user-specific basket ID, which is passed dynamically to the checkout system after we programmatically add a package to the user's basket.

Key notes include:
- Added the Tebex checkout script to the project's public directory to overcome CORS issues temporarily for development and referenced it directly from the root path as specified in the store route's handle.
- Modified the store component to include a conditionally rendered checkout process that initializes only when certain criteria are met, such as the modal being open, the user being authorized, and a valid basket ID with package added being present.
- Implemented a custom React hook, `useTebexCheckout`, which abstracts the checkout initialization logic. This hook takes in the basket ID and the theme ('light' or 'dark') and is triggered based on the component's hydration state and the presence of the Tebex global object.
- Configured the `ExternalScriptsHandle` in the store route to ensure the checkout script is preloaded efficiently when the store route is accessed.
- Included error handling for missing source map files by removing the source map reference from the local script to prevent loading errors and console warnings.

This setup ensures a seamless integration with our existing React/Remix architecture, improving the user's checkout experience by leveraging the Tebex platform's robust e-commerce capabilities directly within our application.