Add Docker Build and Deployment workflow
This commit adds a GitHub Actions workflow file (`build_and_deploy.yml`) for automating the build and deployment process of the Docker image for our project. The workflow is triggered on pushes to the `dev` branch and performs the following tasks:

1. Generates a `.env` file containing the necessary environment variables for the Docker build.
2. Sets up Docker Buildx and logs in to the GitHub Container Registry (GHCR) using Docker login action.
3. Builds and pushes the Docker image to the GitHub Container Registry.
4. Logs out from GHCR after the image is successfully pushed.
5. Logs the completion of the workflow.

The workflow includes error handling to fail the workflow if login to GHCR fails. This automation streamlines our CI/CD pipeline, ensuring consistent and efficient deployment of our application.