Securely inject environment variables at runtime and use Basic Auth
Resolved an authentication issue for Tebex package creation and implemented a secure method for handling environment variables. Removed the creation of a `.env` file from the GitHub Actions workflow to prevent the exposure of sensitive environment variables in a public repository. Environment variables are now securely passed at runtime using GitHub Secrets. The Dockerfile has been left unchanged (just formatted) since environment variables are not expected to be copied into the build; instead, they are injected when the container is instantiated, providing enhanced security and maintaining the confidentiality of credentials.